// app/(shows)/types/show_types.ts
export type Review = {
    userId: string;
    showId: string;
    rating: number;
    movie_title: string;
    movie_poster: string;
    movie_id: string;
    comment: string;
}

export type TvReview = {
    userId: string;
    showId: string;
    rating: number;
    tv_title: string;
    tv_poster: string;
    tv_id: string;
    comment: string;
}